fix: resolve CVE-1989-0001
Verse 1:
There was a glitch in the system
A vulnerability that we couldn't ignore
But we came together and worked it out
And now our code is stronger than before

Chorus:
Cause we resolved the vulnerability
With our teamwork and expertise
We're better than we used to be
And now our system's more secure, can't you see?

Verse 2:
We had to act fast and be nimble
To fix the issue and make it right
But we didn't let it bring us down
We rose up and won the fight

Chorus:
Cause we resolved the vulnerability
With our teamwork and expertise
We're better than we used to be
And now our system's more secure, can't you see?

Bridge:
We know that security's a journey
And we'll keep on improving every day
But for now, we'll celebrate this victory
And keep our code safe from harm's way

Chorus:
Cause we resolved the vulnerability
With our teamwork and expertise
We're better than we used to be
And now our system's more secure, can't you see?

Outro:
So let's raise a glass to our success
And toast to our security team
We'll keep on working hard and doing our best
To keep our system safe and clean.